requirements.txt
pandas
scikit-learn
mlflow
fastapi
uvicorn
joblib
pyyaml